Refactor tooltip out into `FunnelTooltip.tsx`
